requirements.txt
# Vision-Language Model 핵심 패키지
# Python 3.8+ 호환

# PyTorch (CUDA 11.6)
torch==1.12.1
torchvision==0.13.1
torchaudio==0.12.1

# Transformers & NLP
transformers==4.46.3
tokenizers==0.20.3
sentencepiece==0.2.0

# 이미지 처리
pillow==10.4.0

# Hugging Face
huggingface-hub==0.36.0
safetensors==0.5.3
accelerate==1.0.1

# 유틸리티
numpy==1.24.4
tqdm==4.67.1
requests==2.32.4
pyyaml==6.0.3
packaging==25.0
regex==2024.11.6

# 의존성
filelock==3.12.4
sympy==1.13.3
typing-extensions==4.7.1
fsspec==2025.3.0
psutil==7.1.3
protobuf==5.29.5
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
urllib3==2.2.3
mpmath==1.3.0
